requirements.txt
asgiref==3.5.2
astroid==2.12.5
Django==4.1
djangorestframework==3.13.1
djangorestframework-recursive==0.1.2
ffmpeg-python==0.2.0
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
Pillow-SIMD==9.0.0.post1
pylint==2.15.0
pysqlite3-binary==0.4.6
pytz==2022.2.1
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
wrapt==1.14.1